MGDOBR-205: Migrate End2EndTestIT into BDD test
