+++ integration-tests/src/test/resources/application.properties
@@ -1,9 +1,7 @@
 quarkus.http.test-port=8081
-
-event-bridge.manager.url=${EVENT_BRIDGE_MANAGER_URL:http://localhost:8081}
-
+event-bridge.manager.url=http://<REPLACE_WITH_MINIKUBE_IP>:/manager
+key-cloak.url=http://<REPLACE_WITH_MINIKUBE_IP>/auth/realms/event-bridge-fm
 kafka.security.protocol=PLAINTEXT
 kafka.sasl.jaas.config=none
-
 # SSO
 quarkus.oidc.client-id=event-bridge-fm
